add SSH permission fields to Role model and update permissions.py checks
